changed switch443 module so it can handle generic on/off commands
